--- packages/libs/eda/src/lib/notebook/Utils.ts
@@ -1,9 +1,0 @@
-// Utils
-
-export function parseJson(str: string) {
-  try {
-    return JSON.parse(str);
-  } catch {
-    return undefined;
-  }
-}
